project/requirements.txt
PyQt5>=5.15.0
Pillow>=9.0.0
pygame>=2.0.0
